- some code cleanup (more to come) - reduced amount of iterations in password encryption for performance reasons - added checks for every field in installation form to assure that they're properly filled - changed REGON field from int to long as REGON could have 14 digits

+++ SyZaFi/DBConnection.cs
@@ -681,7 +681,7 @@ namespace SyZaFi
             }
         }
 
-        public void InsertNewCompany(string companyName, int companyNIP, int companyREGON, int companyKRS)
+        public void InsertNewCompany(string companyName, int companyNIP, long companyREGON, int companyKRS)
         {
             string query = "INSERT INTO companydetails (companyName, companyNIP, companyREGON, companyKRS)" +
                 " VALUES('" + companyName + "', '" + companyNIP + "', '" + companyREGON + "', '" + companyKRS + "')";
